requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.9.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
av==12.3.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.2
filelock==3.16.0
flatbuffers==24.3.25
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
google-pasta==0.2.0
grpcio==1.67.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.24.6
idna==3.8
ifaddr==0.2.0
imageio==2.35.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.6.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.6.3
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.5
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.40.0
streamlit==1.38.0
streamlit-webrtc==0.47.7
sympy==1.13.2
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2024.8.30
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.32.0
Werkzeug==3.0.4
wrapt==1.16.0
